1.增加 image如果是stb-image，那么使用stb的释放内存函数

+++ include/algorithm.h
@@ -4,7 +4,7 @@
 
 namespace opendip {
 
-    Image LinearInterpolation(Image &src_image, size_t resize_row, size_t resize_col);
+    Image LinearInterpolation(Image &src_image, int resize_row, int resize_col);
 
 }
